chore(db): switch MySQL → PostgreSQL 16
TypeORM-abstracted swap. Driver, container, env defaults, and one
MySQL-only column type (tinyint → smallint) changed; no schema or
query-layer rewrites needed.

Changes:
- backend/src/database/database.module.ts — `type: 'postgres'`; drop
  `charset: utf8mb4` and `timezone: 'Z'` (MySQL-only options)
- backend/src/modules/review/review.entity.ts — rating `tinyint` →
  `smallint` (Postgres has no tinyint; smallint covers 1-5 rating range)
- backend/package.json — replace `mysql2` with `pg` + `@types/pg`
- backend/.env.example — DB_PORT 3306 → 5432, header MySQL → PostgreSQL
- .env.example — drop MYSQL_ROOT_PASSWORD, DB_PORT 3306 → 5432
- docker-compose.yml — `mysql:8.0` → `postgres:16-alpine`; healthcheck
  uses `pg_isready`; volume `mysql_data` → `postgres_data`; api service
  depends_on / DB_HOST / DB_PORT updated
- README.md — stack table says PostgreSQL
- backend/test/e2e/job-lifecycle.e2e-spec.ts — header comment updated

Entity types compatible without further changes:
- `type: 'enum'` → Postgres native enum (TypeORM handles type creation)
- `type: 'json'` → Postgres json (TypeORM maps directly)
- `type: 'decimal'` → numeric (identical semantics)

Verification:
- npm run build ✅
- npm test — 324/324 tests pass across 37 suites; payments coverage 99%
- Live boot smoke test against `postgres:16-alpine` container:
  TypeOrmModule initialized successfully, synchronize ran without error.
  Later boot failure was unrelated (Bull queue needed Redis, expected
  when running outside docker-compose).

+++ backend/src/modules/review/review.entity.ts
@@ -19,7 +19,7 @@ export class Review {
   @Column({ name: 'reviewee_id' })
   revieweeId: string;
 
-  @Column({ type: 'tinyint' })
+  @Column({ type: 'smallint' })
   rating: number;
 
   @Column('text', { nullable: true })
